Extract workspace video settings hook
